+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# [1.7.0](https://github.com/youversion/platform-sdk-kotlin/compare/1.6.1...1.7.0) (2026-07-16)
+
+
+### Features
+
+* **reader:** add line-spacing control to font settings sheet ([cd45d3e](https://github.com/youversion/platform-sdk-kotlin/commit/cd45d3ef4b52c8fd1fb6bdfeaf07fcadcad6f54f))
+
+## [1.6.1](https://github.com/youversion/platform-sdk-kotlin/compare/1.6.0...1.6.1) (2026-07-15)
+
+
+### Bug Fixes
+
+* **ci:** allow bracketed [bot] form of platform-localization-pr-bot in guardrail ([#160](https://github.com/youversion/platform-sdk-kotlin/issues/160)) ([43b04de](https://github.com/youversion/platform-sdk-kotlin/commit/43b04de0dcffb19c65eb55f53e95f4ab5605cb30))
+
 # [1.6.0](https://github.com/youversion/platform-sdk-kotlin/compare/1.5.0...1.6.0) (2026-05-12)